Fix Flash bug, add shooting enemy and bullet physics, uhhhhhhhhhhhhhhhhhhhhhhhhh

+++ include/engine/utilities/rand.h
@@ -1,4 +1,4 @@
 #pragma once
 
 float rand_float(float min, float max);
-float rand_int(int min, int max);
+int rand_int(int min, int max);
