+++ src/engine/utilities/rand.c
@@ -7,6 +7,6 @@ float rand_float(float min, float max) {
 	return ((float)rand() / (float)RAND_MAX) * (max - min) + min;
 }
 
-float rand_int(int min, int max) {
+int rand_int(int min, int max) {
 	return rand() % (max - min) + 1 + min;
 }
